% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.transfer_operations.R
\name{create_user}
\alias{create_user}
\title{Adds a user and associate them with an existing Secure File Transfer Protocol (SFTP) server}
\usage{
create_user(HomeDirectory = NULL, Policy = NULL, Role, ServerId,
  SshPublicKeyBody = NULL, Tags = NULL, UserName)
}
\arguments{
\item{HomeDirectory}{The landing directory (folder) for a user when they log in to the server using their SFTP client. An example is \code{/home/username username }.}

\item{Policy}{A scope-down policy for your user so you can use the same IAM role across multiple users. This policy scopes down user access to portions of their Amazon S3 bucket. Variables you can use inside this policy include \code{${Transfer:UserName}}, \code{${Transfer:HomeDirectory}}, and \code{${Transfer:HomeBucket}}.}

\item{Role}{[required] The IAM role that controls your user's access to your Amazon S3 bucket. The policies attached to this role will determine the level of access you want to provide your users when transferring files into and out of your Amazon S3 bucket or buckets. The IAM role should also contain a trust relationship that allows the SFTP server to access your resources when servicing your SFTP user's transfer requests.}

\item{ServerId}{[required] A system-assigned unique identifier for an SFTP server instance. This is the specific SFTP server that you added your user to.}

\item{SshPublicKeyBody}{The public portion of the Secure Shall (SSH) key used to authenticate the user to the SFTP server.}

\item{Tags}{Key-value pairs that can be used to group and search for users. Tags are metadata attached to users for any purpose.}

\item{UserName}{[required] A unique string that identifies a user and is associated with a server as specified by the \code{ServerId}.}
}
\description{
Adds a user and associate them with an existing Secure File Transfer Protocol (SFTP) server. Using parameters for \code{CreateUser}, you can specify the user name, set the home directory, store the user's public key, and assign the user's AWS Identity and Access Management (IAM) role. You can also optionally add a scope-down policy, and assign metadata with tags that can be used to group and search for users.
}
\details{
The response returns the \code{UserName} and \code{ServerId} values of the new user for that server.
}
\section{Accepted Parameters}{
\preformatted{create_user(
  HomeDirectory = "string",
  Policy = "string",
  Role = "string",
  ServerId = "string",
  SshPublicKeyBody = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  ),
  UserName = "string"
)
}
}

