% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.transfer_operations.R
\name{delete_server}
\alias{delete_server}
\title{Deletes the Secure File Transfer Protocol (SFTP) server that you specify}
\usage{
delete_server(
  ServerId = "string"
)
}
\arguments{
\item{ServerId}{[required] A unique system-assigned identifier for an SFTP server instance.}
}
\description{
Deletes the Secure File Transfer Protocol (SFTP) server that you specify. If you used \code{SERVICE_MANAGED} as your \code{IdentityProviderType}, you need to delete all users associated with this server before deleting the server itself
}
\details{
No response returns from this call.
}
\examples{

}
