% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.transfer_operations.R
\name{delete_ssh_public_key}
\alias{delete_ssh_public_key}
\title{Deletes a user's Secure Shell (SSH) public key}
\usage{
delete_ssh_public_key(ServerId, SshPublicKeyId, UserName)
}
\arguments{
\item{ServerId}{[required] A system-assigned unique identifier for a Secure File Transfer Protocol (SFTP) server instance that has the user assigned to it.}

\item{SshPublicKeyId}{[required] A unique identifier used to reference your user's specific SSH key.}

\item{UserName}{[required] A unique string that identifies a user whose public key is being deleted.}
}
\description{
Deletes a user's Secure Shell (SSH) public key.
}
\details{
No response is returned from this call.
}
\section{Accepted Parameters}{
\preformatted{delete_ssh_public_key(
  ServerId = "string",
  SshPublicKeyId = "string",
  UserName = "string"
)
}
}

