% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.transfer_operations.R
\name{delete_user}
\alias{delete_user}
\title{Deletes the user belonging to the server you specify}
\usage{
delete_user(
  ServerId = "string",
  UserName = "string"
)
}
\arguments{
\item{ServerId}{[required] A system-assigned unique identifier for an SFTP server instance that has the user assigned to it.}

\item{UserName}{[required] A unique string that identifies a user that is being deleted from the server.}
}
\description{
Deletes the user belonging to the server you specify.
}
\details{
No response returns from this call.

When you delete a user from a server, the user's information is lost.
}
\examples{

}
