% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.transfer_operations.R
\name{describe_server}
\alias{describe_server}
\title{Describes the server that you specify by passing the \code{ServerId} parameter}
\usage{
describe_server(
  ServerId = "string"
)
}
\arguments{
\item{ServerId}{[required] A system-assigned unique identifier for an SFTP server.}
}
\description{
Describes the server that you specify by passing the \code{ServerId} parameter.
}
\details{
The response contains a description of the server's properties.
}
\examples{

}
