% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.transfer_operations.R
\name{describe_user}
\alias{describe_user}
\title{Describes the user assigned to a specific server, as identified by its \code{ServerId} property}
\usage{
describe_user(
  ServerId = "string",
  UserName = "string"
)
}
\arguments{
\item{ServerId}{[required] A system-assigned unique identifier for an SFTP server that has this user assigned.}

\item{UserName}{[required] The name of the user assigned to one or more servers. User names are part of the sign-in credentials to use the AWS Transfer service and perform file transfer tasks.}
}
\description{
Describes the user assigned to a specific server, as identified by its \code{ServerId} property.
}
\details{
The response from this call returns the properties of the user associated with the \code{ServerId} value that was specified.
}
\examples{

}
