% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.transfer_operations.R
\name{import_ssh_public_key}
\alias{import_ssh_public_key}
\title{Adds a Secure Shell (SSH) public key to a user account identified by a \code{UserName} value assigned to a specific server, identified by \code{ServerId}}
\usage{
import_ssh_public_key(
  ServerId = "string",
  SshPublicKeyBody = "string",
  UserName = "string"
)
}
\arguments{
\item{ServerId}{[required] A system-assigned unique identifier for an SFTP server.}

\item{SshPublicKeyBody}{[required] The public key portion of an SSH key pair.}

\item{UserName}{[required] The name of the user account that is assigned to one or more servers.}
}
\description{
Adds a Secure Shell (SSH) public key to a user account identified by a \code{UserName} value assigned to a specific server, identified by \code{ServerId}.
}
\details{
The response returns the \code{UserName} value, the \code{ServerId} value, and the name of the \code{SshPublicKeyId}.
}
\examples{

}
