% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.transfer_operations.R
\name{list_servers}
\alias{list_servers}
\title{Lists the Secure File Transfer Protocol (SFTP) servers that are associated with your AWS account}
\usage{
list_servers(MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{MaxResults}{Specifies the number of servers to return as a response to the \code{ListServers} query.}

\item{NextToken}{When additional results are obtained from the ListServers command, a \code{NextToken} parameter is returned in the output. You can then pass the \code{NextToken} parameter in a subsequent command to continue listing additional servers.}
}
\description{
Lists the Secure File Transfer Protocol (SFTP) servers that are associated with your AWS account.
}
\section{Accepted Parameters}{
\preformatted{list_servers(
  MaxResults = 123,
  NextToken = "string"
)
}
}

