% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.transfer_package.R
\docType{package}
\name{paws.transfer-package}
\alias{paws.transfer}
\alias{paws.transfer-package}
\title{paws.transfer: AWS Transfer for SFTP}
\description{

}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
