% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.transfer_operations.R
\name{start_server}
\alias{start_server}
\title{Changes the state of a Secure File Transfer Protocol (SFTP) server from \code{OFFLINE} to \code{ONLINE}}
\usage{
start_server(
  ServerId = "string"
)
}
\arguments{
\item{ServerId}{[required] A system-assigned unique identifier for an SFTP server that you start.}
}
\description{
Changes the state of a Secure File Transfer Protocol (SFTP) server from \code{OFFLINE} to \code{ONLINE}. It has no impact on an SFTP server that is already \code{ONLINE}. An \code{ONLINE} server can accept and process file transfer jobs.
}
\details{
The state of \code{STARTING} indicates that the server is in an intermediate state, either not fully able to respond, or not fully online. The values of \code{START_FAILED} can indicate an error condition.

No response is returned from this call.
}
\examples{

}
