% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.transfer_operations.R
\name{stop_server}
\alias{stop_server}
\title{Changes the state of an SFTP server from ONLINE to OFFLINE}
\usage{
stop_server(ServerId)
}
\arguments{
\item{ServerId}{[required] A system-assigned unique identifier for an SFTP server that you stopped.}
}
\description{
Changes the state of an SFTP server from \code{ONLINE} to \code{OFFLINE}. An \code{OFFLINE} server cannot accept and process file transfer jobs. Information tied to your server such as server and user properties are not affected by stopping your server. Stopping a server will not reduce or impact your Secure File Transfer Protocol (SFTP) endpoint billing.
}
\details{
The states of \code{STOPPING} indicates that the server is in an intermediate state, either not fully able to respond, or not fully offline. The values of \code{STOP_FAILED} can indicate an error condition.

No response is returned from this call.
}
\section{Accepted Parameters}{
\preformatted{stop_server(
  ServerId = "string"
)
}
}

