% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.transfer_operations.R
\name{update_server}
\alias{update_server}
\title{Updates the server properties after that server has been created}
\usage{
update_server(
  IdentityProviderDetails = list(
    Url = "string",
    InvocationRole = "string"
  ),
  LoggingRole = "string",
  ServerId = "string"
)
}
\arguments{
\item{IdentityProviderDetails}{This response parameter is an array containing all of the information required to call a customer's authentication API method.}

\item{LoggingRole}{Changes the AWS Identity and Access Management (IAM) role that allows Amazon S3 events to be logged in Amazon CloudWatch, turning logging on or off.}

\item{ServerId}{[required] A system-assigned unique identifier for an SFTP server instance that the user account is assigned to.}
}
\description{
Updates the server properties after that server has been created.
}
\details{
The \code{UpdateServer} call returns the \code{ServerId} of the Secure File Transfer Protocol (SFTP) server you updated.
}
\examples{

}
